+++ src/vscode/extension.ts
@@ -98,22 +98,28 @@ class FactorioDebugProvider implements vscode.DebugConfigurationProvider, vscode
 		const activeVersion = await this.versionSelector.getActiveVersion();
 		if (!activeVersion) { return; }
 
-		if (!config.factorioArgs) {
-			config.factorioArgs = [];
+		const factorioArgs = config.factorioArgs??[] as string[];
+
+		if (activeVersion.configPathIsOverriden()) {
+			if (!factorioArgs.includes("-c") && !factorioArgs.includes("--config")) {
+				factorioArgs.push( "--config", await activeVersion.configPath());
+			}
 		}
 
-		if (!(config.factorioArgs as string[]).includes("--mod-directory")) {
+		if (!factorioArgs.includes("--mod-directory")) {
 			const modlists = await vscode.workspace.findFiles("**/mod-list.json");
 			if (modlists.length === 1) {
 				const dir = path.posix.normalize(path.resolve(vscode.Uri.joinPath(modlists[0], "..").fsPath));
 				const def = await activeVersion.defaultModsPath();
 				if (dir !== def) {
-					config.factorioArgs = [ ...config.factorioArgs, "--mod-directory", dir];
+					factorioArgs.push("--mod-directory", dir);
 				}
 			} else if (modlists.length > 1) {
 				throw new Error("Cannot automatically select mod directory, multiple mod-list.json in workspace");
 			}
 		}
+
+		config.factorioArgs = factorioArgs;
 
 		const debugconfigenv = vscode.workspace.getConfiguration("factorio.debug").get("env", {});
 		config.env = {...debugconfigenv, ...config.env, SteamAppId: "427520"};