+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 [<img height='36' style='border:0px;height:36px;' src='https://az743702.vo.msecnd.net/cdn/kofi2.png?v=2' border='0' alt='Buy Me a Coffee at ko-fi.com'/>](https://ko-fi.com/X8X41IE4T)
 
 ## 2.1.10
+* Debugger:
+  * automatically pass `configPath` from version info as `--config` unless it would conflict
 
 ## 2.1.9
 * Lua Library package